+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 Note, that only significant changes are listed.
 
+## DEV
+
+* no more protocol-agnostic links ('//themes...'); make them explicit https in production
+
+
 ## 0.5.3 (2012-12-20)
 
 * new `budget_host`, `budget_uri` and `budget_require_ssl` helpers